refactor : 서비스 배지 로직을 공통 컴포넌트로 정리
