+++ src/pages/workflows/model/types.ts
@@ -1,20 +1,6 @@
+import type { ServiceBadgeKey } from "@/shared";
+
 import { WORKFLOW_FILTERS } from "./constants";
 
 export type WorkflowFilterKey = (typeof WORKFLOW_FILTERS)[number]["key"];
-
-export type ServiceBadgeKey =
-  | "calendar"
-  | "gmail"
-  | "google-drive"
-  | "google-sheets"
-  | "notion"
-  | "slack"
-  | "communication"
-  | "storage"
-  | "spreadsheet"
-  | "web-scraping"
-  | "notification"
-  | "llm"
-  | "trigger"
-  | "processing"
-  | "unknown";
+export type { ServiceBadgeKey };
