+++ src/shared/index.ts
@@ -4,4 +4,5 @@ export * from "./libs";
 export * from "./model";
 export * from "./theme";
 export * from "./types";
+export * from "./ui";
 export * from "./utils";
