+++ src/shared/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./relative-time";
+export * from "./service-badge";
